more

+++ todo/views.py
@@ -20,22 +20,26 @@ def taskView(request):
             task = Task(title=items["title"], task_text=items["text"], user=user)
             task.save()
         else:
-            tasks = Task.objects.all()
-            for task in tasks:
-                task.status = False
-                task.save()
-            for key, val in items.items():
-                try:
-                    key = int(key)
-                    task = Task.objects.get(id=key)
-                    if val == 'on':
-                        task.status = True
-                        task.save()
-                except ValueError:
-                    if 'delete' in key:
-                        id = int(key.split('delete')[1])
-                        task = Task.objects.get(id=id)
-                        task.delete()
+            if items.get('search') is not None:
+                tasks = Task.objects.filter(user=user.id, title__contains=items["search"])
+                return render(request, "task/tasks.html", context={"tasks": tasks})
+            else:
+                tasks = Task.objects.all()
+                for task in tasks:
+                    task.status = False
+                    task.save()
+                for key, val in items.items():
+                    try:
+                        key = int(key)
+                        task = Task.objects.get(id=key)
+                        if val == 'on':
+                            task.status = True
+                            task.save()
+                    except ValueError:
+                        if 'delete' in key:
+                            id = int(key.split('delete')[1])
+                            task = Task.objects.get(id=id)
+                            task.delete()
         return redirect('/')
     tasks = Task.objects.filter(user=user.id)
     return render(request, "task/tasks.html", context={"tasks": tasks})
